Release 1.2.0

+++ apiserver/build.sh
@@ -2,7 +2,7 @@
 
 BUILD_DATE=`date +%Y-%m-%d\ %H:%M`
 VERSIONFILE="pkg/version/version.go"
-VERSION="1.1.0"
+VERSION="1.2.0"
 
 
 if [ "$1" == "local" ] || [ "$1" == "docker" ]; then
@@ -17,7 +17,13 @@ if [ "$1" == "local" ] || [ "$1" == "docker" ]; then
     echo "  BUILD_DATE = \"$BUILD_DATE\"" >> $VERSIONFILE
     echo ")" >> $VERSIONFILE
     
-    glide install --strip-vendor 
+    # Check for --cache flag
+    args="$@"
+    replaced="${@/--cache/}"
+    if [ "$1" == "local" ] && [ "$args" == "$replaced" ]; then
+        echo "Fetching dependencies..."
+        glide install --strip-vendor
+    fi
     
     COVERPKG=./cmd/server,./pkg/crypto,./pkg/etcd,./pkg/config,./pkg/email,./pkg/events,./pkg/kube,./pkg/middleware,./pkg/types,./pkg/validate
 	if [ "$1" == "local" ]; then 
@@ -57,7 +63,6 @@ if [ "$1" == "local" ] || [ "$1" == "docker" ]; then
           GOOS=darwin GOARCH=amd64 go build -o build/bin/ndslabsctl-darwin-amd64 ./cmd/apictl
     fi
     
-    rm -r pkg/version
 elif [ "$1" == "clean" ]; then
 	rm -r build
 	rm -r vendor/github.com vendor/golang.org vendor/gopkg.in vendor/k8s.io
